Refactor to improve consistency
