+++ jodeln/network/net.py
@@ -2,231 +2,20 @@
 """
 
 import sys
-from dataclasses import dataclass
-from .geh import geh
-from typing import List, Dict, Tuple, Generator
 from collections import Counter
+from typing import TYPE_CHECKING, Generator
 
+from .geh import geh
+from .netlink import NetLinkData
+from .netnode import NetNode
+from .netod import NetODpair
+from .netroute import NetRoute
+from .netturns import TurnData
 
-@dataclass
-class NetRoute():
-    """Contains sequence of nodes from origin to destination.
-
-    Attributes
-    ----------
-    nodes : List[int]
-        Ordered sequence of node IDs along the route, from origin to destination.
-    name : str
-        Human-readable name for the route. Does not have to be unique amongst all routes.
-    seed_volume : float
-        Volume on the route assigned from the seed OD matrix.
-    target_ratio : float
-        If one OD pair has multiple routes, this is the desired proportion of the 
-        total OD volume that should occur on this route.
-    target_rel_diff : float
-        If one OD pair has multiple routes, this is a helper variable for the OD
-        optimization that indicates how different this route is relative to all 
-        other routes within the OD. For example, 2 routes, one has a target_ratio 
-        of 0.8, the other 0.20. The target relative differences are 0.60 and -0.60, 
-        respectively.
-    assigned_volume : float
-        Volume assigned to this route based on the estimated OD matrix.
-    assigned_ratio : float
-        If one OD pair has multiple routes, this is the actual proportion of the 
-        total estimated OD volume. assigned_ratio is compared to the target_ratio
-        in the OD optimization.
-    opt_var_index : int
-        Helper variable for the OD optimization algorithm. The OD optimization
-        takes a list of variables. opt_var_index is the position of this route 
-        within the list of optimization variables.
-    """
-    __slots__ = ['nodes', 'name', 'seed_volume', 'target_ratio', 'target_rel_diff',
-                 'assigned_volume', 'assigned_ratio', 'opt_var_index']
-    nodes: List[int]
-    name: str
-    seed_volume: float
-    target_ratio: float
-    target_rel_diff: float
-    assigned_volume: float
-    assigned_ratio: float
-    opt_var_index: int
-
-
-@dataclass
-class NetODpair():
-    """Origin-Destination data in the Network. 
+if TYPE_CHECKING:
+    from .netlink import LinkParameters
+    from .netnode import NodeParameters
     
-    Attributes
-    ----------
-    origin : int
-        node ID of the OD origin.
-    destination : int
-        node ID of the OD destination.
-    seed_total_volume : float
-        Volume for this OD pair within the seed OD matrix.
-    est_total_volume : float
-        Volume for this OD pair within the estimated OD matrix.
-    routes : List[route]
-        List of all possible routes from origin to destination to include in the
-        OD optimization. One OD can contain multiple routes.
-    """
-    __slots__ = ['origin', 'destination', 'seed_total_volume', 'est_total_volume', 'routes']
-    origin: int
-    destination: int
-    seed_total_volume: float
-    est_total_volume: float
-    routes: List[NetRoute]
-
-@dataclass
-class LinkParameters():
-    """Parameters needed for constructing a NetLinkData object.
-    """
-    __slots__ = ['name', 'cost', 'target_volume', 'shape_points']
-    name: str
-    cost: float
-    target_volume: float
-    shape_points: List[Tuple[float, float]]
-
-
-@dataclass
-class NetLinkData():
-    """Data on Network links.
-
-    Attributes
-    ----------
-    cost: float
-        Friction on the link that is used when determining shortest routes.
-    name: str
-        Human-readable name for the link. Does not have to be unique.
-    target_volume: float
-        Desired volume on this link.
-    link_index: int
-        Unique identifier for this link.
-    assigned_volume: float
-        Volume on the link as assigned from the estimated OD matrix.
-    geh: float
-        GEH statistic comparing the target_volume and assigned_volume.
-    seed_volume: float
-        Volume on the link as assigned from the seed OD matrix.
-    """
-    __slots__ = ['cost', 'name', 'target_volume', 'link_index', 'assigned_volume', 'geh', 'seed_volume', 'shape_points']
-    cost: float
-    name: str
-    target_volume: float
-    link_index: int
-    assigned_volume: float
-    geh: float
-    seed_volume: float
-    shape_points: List[Tuple[float, float]]
-
-
-@dataclass
-class TurnData():
-    """Data on Network turns.
-    
-    A turn is a sequence of three consecutive nodes: A-B-C. In this case "B"
-    is the intersection, A is upstream, and C is downstream.
-
-    Attributes
-    ----------
-    key: int
-        Unique ID for the turn.
-    name: str
-        Human-readable name for the turn. Typically based on the A-B-C node names.
-    seed_volume: float
-        Volume on the turn as assigned from the seed OD matrix.
-    target_volume: float
-        Desired volume on this turn.
-    assigned_volume: float
-        Volume on the turn as assigned from the estimated OD matrix.
-    geh: float
-        GEH statistic comparing the target_volume and assigned_volume.
-    """
-    __slots__ = ['key', 'name', 'seed_volume', 'target_volume', 'assigned_volume', 'geh']
-    key: int
-    name: str
-    seed_volume: float
-    target_volume: float
-    assigned_volume: float
-    geh: float
-
-
-@dataclass
-class NodeParameters():
-    """Parameters needed for constructing a NetNode object.
-    """
-    __slots__ = ['name', 'x', 'y', 'is_origin', 'is_destination']
-    name: str
-    x: float
-    y: float
-    is_origin: bool
-    is_destination: bool
-
-
-class NetNode():
-    """A point (junction) in the network graph.
-
-    Two connected nodes form an link. A sequence of three nodes forms a 
-    Turn (see TurnData).
-
-    Attributes
-    ----------
-    key : int
-        Unique identifier for the node.
-    name : str
-        Name, or label, for the node.
-    x : float
-        x-coordinate of node. Defaults to zero.
-    y : float
-        y-coordinate of node. Defaults to zero.
-    is_origin : bool
-        Indicates if traffic can start their trip from this node (source node)
-    is_destination : bool
-        Indicates if traffic can end their trip at this node (sink node)
-    neighbors : Dict[int, NetLinkData]
-        Adjacency list of connected downstream node IDs and associated data
-    up_neighbors : list
-        Upstream node IDs connected to this node.
-    """
-    def __init__(self, key, parameters: NodeParameters):
-        """Create a node in the network graph.
-
-        Parameters
-        ----------
-        key : int
-            Unique identifier.
-        parameters : NodeParameters
-            Data about the node. Name, x,y coordinates, etc.
-        """
-        self.key = key
-        
-        self.name = parameters.name
-        self.x = parameters.x
-        self.y = parameters.y
-        self.is_origin = parameters.is_origin
-        self.is_destination = parameters.is_destination
-
-        self.neighbors = {} # type: Dict[int, NetLinkData]
-        self.up_neighbors = []
-
-    def add_neighbor(self, neighbor, link_data) -> None:
-        """Connects two nodes to form an link.
-
-        links are stored as an adjacency list.
-
-        Parameters
-        ----------
-        neighbor : int
-            Unique identifier of neighboring (connecting) node.
-        link_data : NetLinkData
-            See NetLinkData class.
-        """
-        self.neighbors[neighbor] = link_data
-
-    def get_connections(self):
-        """Return list of node keys connected to self node."""
-        return self.neighbors.keys()
-
 
 class Network():
     """Contains the network nodes and links, turns, and assigned origin-destination 
@@ -238,7 +27,7 @@ class Network():
 
     Attributes
     ----------
-    nodes : Dict[int, NetNode]
+    _graph : Dict[int, NetNode]
         nodes within the Network graph.
     turns : Dict[int, TurnData]
         Turns within the Network graph.
@@ -246,24 +35,24 @@ class Network():
         number of links in the network.
     od : List[NetODpair]
         OD data for the network.
-    total_geh : int
+    total_geh : float
         Grand total of summing all the GEH values of the links and turns. 
         See the calc_network_geh method.
     coord_scale : float
         Scalar to convert node x,y position to real-world coordinates. Required
         to ensure the network is displayed legibly in the GUI.
     """
-    __slots__ = ['nodes', 'turns', 'n_links', 'od', 'total_geh', 'coord_scale']
+    __slots__ = ['_graph', '_turns', 'n_links', 'od', 'total_geh', 'coord_scale']
 
     def __init__(self):
-        self.nodes = {} # type: Dict[int, NetNode]
-        self.turns = {}  # type: Dict[Tuple[int, int, int], TurnData]
-        self.n_links = 0
-        self.od = [] # type: List[NetODpair]
-        self.total_geh = 0
-        self.coord_scale = 1
+        self._graph: dict[int, NetNode] = {}
+        self._turns: dict[tuple[int, int, int], TurnData] = {}
+        self.n_links: int = 0
+        self.od: list[NetODpair] = []
+        self.total_geh: float = 0
+        self.coord_scale: float = 1
 
-    def add_node(self, parameters: NodeParameters) -> None:
+    def add_node(self, parameters: 'NodeParameters') -> None:
         """Add a node to the network graph.
 
         Parameters
@@ -271,10 +60,11 @@ class Network():
         parameters : NodeParameters
             Data about the node. Name, x,y coordinates, etc.
         """
-        key = len(self.nodes)
-        self.nodes[key] = NetNode(key, parameters)
+        # FIXME: length not guaranteed to return a unique key number.
+        key = len(self._graph)
+        self._graph[key] = NetNode(key, parameters)
 
-    def add_link(self, i_name, j_name, parameters: LinkParameters) -> None:
+    def add_link(self, i_name, j_name, parameters: 'LinkParameters') -> None:
         """Connects two nodes to form an link in the network graph.
 
         Parameters
@@ -299,64 +89,81 @@ class Network():
             seed_volume=0,
             shape_points=parameters.shape_points)
 
-        self.nodes[i_key].add_neighbor(j_key, link_data)
+        self._graph[i_key].add_neighbor(j_key, link_data)
         self.n_links += 1
 
-        self.nodes[j_key].up_neighbors.append(i_key)
-        
-    def link(self, i, j):
+        self._graph[j_key].up_neighbors.append(i_key)
+
+    def node(self, key: int) -> NetNode:
+        """Convenience function to access node properties."""
+        # TODO: Handle case if key is not in _graph.
+        return self._graph[key]
+
+    def nodes(self, return_keys: bool=False) -> \
+        Generator[NetNode, None, None] | \
+        Generator[tuple[int, NetNode], None, None]:
+        """Generator function to return all network nodes"""
+        for key, node in self._graph.items():
+            if return_keys:
+                yield key, node
+            else:
+                yield node
+
+    def link(self, i: int, j: int) -> NetLinkData:
         """Convenience function to access link properties."""
-        return self.nodes[i].neighbors[j]
+        return self._graph[i].neighbors[j]
 
-    def links_(self) -> Generator[Tuple[int, int, NetLinkData], None, None]:
-        """Generator function to iterate through all the links.
-        
-        Function name has a trailing underscore for consistency with self.turns_()
-        """
-        for i, node in self.nodes.items():
+    def links(self, return_keys: bool=False) -> \
+        Generator[NetLinkData, None, None] | \
+        Generator[tuple[tuple[int, int], NetLinkData], None, None]:
+        """Generator function to iterate through all the links."""
+        for i, node in self._graph.items():
             for j, _ in node.neighbors.items():
-                yield (i, j, self.nodes[i].neighbors[j])
+                if return_keys:
+                    yield (i, j), self._graph[i].neighbors[j]
+                else:
+                    yield self._graph[i].neighbors[j]
 
-    def turns_(self) -> Generator[Tuple[Tuple[int, int, int], TurnData], None, None]:
-        """Generator function to itertae through all the turns.
+    def turn(self, i: int, j: int, k: int) -> TurnData:
+        """Convenience function to access turn properties."""
+        return self._turns[(i, j, k)]
 
-        Function name has a trailing underscore to avoid conflict with self.turns
-        variable name. Function is implemented to create consistent looking code 
-        when iterating through all the turns or iterating through all the links.
-        e.x.:
-           "for k, v in net.turns_()" or "for k, v in net.links_()"
-        """
-        for (i, j, k), turn in self.turns.items():
-            yield (i, j, k), turn
+    def turns(self, return_keys: bool=False) -> \
+        Generator[TurnData, None, None] | \
+        Generator[tuple[tuple[int, int, int], TurnData], None, None]:
+        """Generator function to itertate through all the turns."""
+        for key, turn in self._turns.items():
+            if return_keys:
+                yield key, turn
+            else:
+                yield turn
 
 
     def init_turns(self) -> None:
         """Initialize all turns within the network."""
 
-        turn_counter = 0
-        for i, node1 in self.nodes.items():
+        for i, node1 in self._graph.items():
             for j, _ in node1.neighbors.items():
-                for k, _ in self.nodes[j].neighbors.items():
-                    self.turns[(i, j, k)] = TurnData(key=turn_counter,
+                for k, _ in self._graph[j].neighbors.items():
+                    self._turns[(i, j, k)] = TurnData(key=(i, j, k),
                                                      name=f'{i}_{j}_{k}',
                                                      seed_volume=0,
                                                      target_volume=0,
                                                      assigned_volume=0,
                                                      geh=0)
-                    turn_counter += 1
 
     def init_routes(self) -> None:
         """Initialize routes by determining shortest route from all origins
         to all destinations."""
 
-        for i, o_node in self.nodes.items():
+        for i, o_node in self._graph.items():
             if not o_node.is_origin:
                 continue
 
             result = _dijkstra(self, i)
 
             # for each destination, get route from O to D
-            for j, d_node in self.nodes.items():
+            for j, d_node in self._graph.items():
                 if not d_node.is_destination:
                     continue
                 
@@ -378,7 +185,7 @@ class Network():
 
     def get_node_by_name(self, node_name):
         """Helper function to return a node by name."""
-        for k, node in self.nodes.items():
+        for k, node in self._graph.items():
             if node.name == node_name:
                 return k, node
         else:
@@ -392,14 +199,14 @@ class Network():
         self.total_geh = 0
         
         # calc link geh
-        for _, _, link in self.links_(): 
+        for link in self.links(): 
             # TODO: handle case when link has no raw volume
             link_geh = geh(link.target_volume, link.assigned_volume)
             link.geh = link_geh
             self.total_geh += link_geh
         
         # calc turn geh
-        for _, t in self.turns_():
+        for t in self.turns():
             # TODO: better handling when turn has no target volume
             if t.target_volume <= 0:
                 continue
@@ -424,19 +231,19 @@ class Network():
 
         self.set_link_and_turn_volume_from_route()
 
-        for _, _, link in self.links_():
+        for link in self.links():
             link.seed_volume = link.assigned_volume
         
-        for _, t in self.turns_():
+        for t in self.turns():
             t.seed_volume = t.assigned_volume
     
     def set_link_and_turn_volume_from_route(self) -> None:
         """Calculate the volume on all links and turns based on the OD route volumes."""
         # reset link & turn volumes to zero
-        for _, _, link in self.links_():
+        for link in self.links():
             link.assigned_volume = 0
         
-        for _, t in self.turns_():
+        for t in self.turns():
             t.assigned_volume = 0
         
         # assign link & turn volumes
@@ -461,7 +268,7 @@ class Network():
                     j = route.nodes[x + 1]
                     k = route.nodes[x + 2]
                     self.link(i, j).assigned_volume += route.assigned_volume
-                    self.turns[(i, j, k)].assigned_volume += route.assigned_volume
+                    self.turn(i, j, k).assigned_volume += route.assigned_volume
 
                 j = route.nodes[x + 1]
                 k = route.nodes[x + 2]
@@ -481,8 +288,8 @@ class Network():
             od_links = []
             for route in od.routes:
                 for x in range(0, len(route.nodes) - 1):
-                    a = self.nodes[route.nodes[x]].name
-                    b = self.nodes[route.nodes[x + 1]].name
+                    a = self._graph[route.nodes[x]].name
+                    b = self._graph[route.nodes[x + 1]].name
                     od_links.append((a, b))
             
             # Find unique links based on counting how many times each link is used
@@ -493,8 +300,8 @@ class Network():
             # Assign route names based on a unique link along the route.
             for route in od.routes:
                 for x in range(0, len(route.nodes) - 1):
-                    a = self.nodes[route.nodes[x]].name
-                    b = self.nodes[route.nodes[x + 1]].name
+                    a = self._graph[route.nodes[x]].name
+                    b = self._graph[route.nodes[x + 1]].name
                     if (a, b) in unique_links:
                         route.name = str(a) + "_" + str(b)
                         # Remove link to prevent assigning the same name to mulitple routes.
@@ -506,7 +313,7 @@ class Network():
         legibly in the GUI. Scale value is saved in self.coord_scale
         """
         coords = []
-        for _, node in self.nodes.items():
+        for node in self.nodes():
             coords.append((node.x, node.y))
         
         # Calculate extents, min/max on x-axis and y-axis
@@ -524,16 +331,16 @@ class Network():
         self.coord_scale = max(scale_factor_x, scale_factor_y)
 
         # Scale node coordinates and link shape points
-        for _, node in self.nodes.items():
+        for node in self.nodes():
             node.x *= self.coord_scale
             node.y *= self.coord_scale
 
-        for _, _, link in self.links_():
+        for link in self.links():
             new_shape_points = [(x * self.coord_scale, y * self.coord_scale) for x, y in link.shape_points]
             link.shape_points = new_shape_points
 
 
-def _dijkstra(net: Network, source):
+def _dijkstra(net: Network, source: int):
     """Uses dijkstra's algorithm to compute the shortest route between
     source and all destinations.
     
@@ -563,7 +370,7 @@ def _dijkstra(net: Network, source):
     # previous node on shortest route
     prev = {}
 
-    for i in net.nodes:
+    for i in net._graph:
         Q.append(i)  
         dist[i] = sys.maxsize
         prev[i] = None
@@ -577,7 +384,7 @@ def _dijkstra(net: Network, source):
 
         Q.remove(u)
 
-        for v, link in net.nodes[u].neighbors.items():
+        for v, link in net._graph[u].neighbors.items():
             alt = dist[u] + link.cost
             if alt < dist[v]:   
                 dist[v] = alt
